Remove tests package (#12)
This commit is to clean up the previous tests package. This package contained modules that were used to
come up with designs on different workflows for using AWX. Some of these modules have been migrated to the demo folder at the root of the project. For all examples on how to use this project, please see the demo folder for more details.